+++ packages/map-template/CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.96.1] - 2026-03-06
 
 ### Added
-- Full screen button configurable via App Config `enableFullScreenButton` under `appSettings`
+- Full screen button (desktop only). Configurable via App Config `enableFullScreenButton` under `appSettings`. Can also be excluded via `excludeFromUI`.
 
 ## [1.96.0] - 2026-03-05